Factor out common parsing and formatting operations to lib/tools
